Moved the files around into their logical folders
