+++ 06_cutVideoIntoFrames.py
@@ -5,12 +5,11 @@ from lib.CommandLineLauncher import CommandLineLauncher
 from lib.FolderStructure import FolderStructure
 from lib.infra.Configurations import Configurations
 from lib.ui.FileOpenUI import FileOpenUI
-from lib.ui.StreamToLogger import StreamToLogger
 from lib.VideoToImages import VideoToImages
 from lib.infra.Logger import Logger
 from lib.VideoStream import VideoStream
 from lib.data.CrabsData import CrabsData
-from lib.data.SeeFloor import SeeFloor
+from lib.seefloor.SeeFloor import SeeFloor
 from lib.seefloor.InterpolateController import InterpolateController
 
 print ("Launched markCrabs script")
