+++ lib/imageProcessing/Rectificator.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from lib.Image import Image
 from lib.Camera import Camera
-from lib.common import Point
+from lib.model.Point import Point
 from lib.VideoStream import VideoStream
 from lib.imageProcessing.EuclidianPlane import EuclidianPlane
 from lib.imageProcessing.PointDetector import PointDetector
