+++ lib/reddots/RedDot.py
@@ -3,7 +3,8 @@ import numpy as np
 from skimage import measure
 
 from lib.Image import Image
-from lib.common import Point, Box
+from lib.model.Box import Box
+from lib.model.Point import Point
 
 
 class RedDot:
