+++ tests/test_DriftManualData.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from lib.FolderStructure import FolderStructure
-from lib.common import Point
+from lib.model.Point import Point
 from lib.drifts.DriftManualData import DriftManualData
 import pandas as pd
