+++ tests/test_Feature.py
@@ -1,8 +1,5 @@
 from unittest import TestCase
 
-from lib.Feature import Feature
-from lib.common import Point
-import pandas as pd
 
 class TestFeature(TestCase):
