+++ tests/test_FramePhysics.py
@@ -4,7 +4,8 @@ from unittest import TestCase
 from lib.Camera import Camera
 from lib.Frame import Frame
 from lib.FramePhysics import FramePhysics
-from lib.common import Vector, Point
+from lib.model.Vector import Vector
+from lib.model.Point import Point
 
 
 class TestFramePhysics(TestCase):
