+++ tests/test_seeFloorWithBadBlocks.py
@@ -6,9 +6,7 @@ from lib.Camera import Camera
 from lib.Frame import Frame
 from lib.data.BadFramesData import BadFramesData
 from lib.data.DriftData import DriftData
-from lib.VideoToImages import VideoToImages
-from lib.data.SeeFloorWithBadBlocks import SeeFloorWithBadBlocks
-from lib.data.SeeFloor import SeeFloor
+from lib.seefloor.SeeFloorWithBadBlocks import SeeFloorWithBadBlocks
 from lib.infra.DataframeWrapper import DataframeWrapper
